+++ react/FormComponent.tsx
@@ -18,7 +18,7 @@ import {
 } from 'vtex.styleguide'
 import { useRuntime } from 'vtex.render-runtime'
 import { useQuery, useMutation } from 'react-apollo'
-import type { InjectedIntlProps } from 'react-intl'
+import type { WrappedComponentProps } from 'react-intl'
 import { injectIntl } from 'react-intl'
 import { formatIOMessage } from 'vtex.native-types'
 
@@ -44,7 +44,9 @@ interface FormComponentProps {
 
 const messages = messagesForm
 
-const FormComponent: FC<FormComponentProps & InjectedIntlProps> = (props) => {
+const FormComponent: FC<FormComponentProps & WrappedComponentProps> = (
+  props
+) => {
   const dataMenuTypeArray: DataMenu[] = []
   const dataMenuType: DataMenu = {
     icon: '',
@@ -429,6 +431,7 @@ const FormComponent: FC<FormComponentProps & InjectedIntlProps> = (props) => {
       case 'optionalText':
         setOptionalText(e.value)
         break
+
       case 'linkBanner':
         setLinkBanner(e.value)
         break
@@ -815,6 +818,7 @@ const FormComponent: FC<FormComponentProps & InjectedIntlProps> = (props) => {
           tempThird[0].optionalText = optionalText
         }
       }
+
       insertSubMenu(
         {
           id: menu.id,